build: update migrations

+++ migrations/20220120083348_create_resources.sql
@@ -10,7 +10,7 @@ CREATE TABLE resources
     name VARCHAR(100) NOT NULL,
     config JSON NOT NULL,
     is_published BOOLEAN,
-    last_modified DOUBLE NOT NULL,
+    last_modified DOUBLE PRECISION NOT NULL,
     last_modified_by VARCHAR(100) NOT NULL,
     message VARCHAR(200) NOT NULL,
     op ENUM('ADDED', 'UPDATED', 'DELETED') NOT NULL,
